requirements.txt
Flask==3.1.0
python-dotenv==1.1.0
psycopg2-binary==2.9.10
flasgger==0.9.1